Wifi: qca6174: Update the wifi driver to 4.5.25.57
To fix system hang problem in the SAP mode, sync the
qcacld driver to the latest version.

+++ drivers/net/wireless/qcacld-2.0/CORE/CLD_TXRX/TXRX/ol_tx.c
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2011-2014, 2016-2018 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2011-2014, 2016-2018, 2021 The Linux Foundation. 
+ * All rights reserved.
  *
  * Previously licensed under the ISC license by Qualcomm Atheros, Inc.
  *
@@ -1987,7 +1988,7 @@ ol_tx_pdev_reset_bundle_require(void* pdev_handle)
  */
 void
 ol_tx_vdev_set_bundle_require(uint8_t vdev_id, unsigned long tx_bytes,
-			uint32_t time_in_ms, uint32_t high_th, uint32_t low_th)
+			uint32_t time_in_ms, uint64_t high_th, uint64_t low_th)
 {
 	struct ol_txrx_vdev_t* vdev = ol_txrx_get_vdev_from_vdev_id(vdev_id);
 	bool old_bundle_required;
@@ -2003,7 +2004,7 @@ ol_tx_vdev_set_bundle_require(uint8_t vdev_id, unsigned long tx_bytes,
 
 	if (old_bundle_required != vdev->bundling_reqired)
 		TXRX_PRINT(TXRX_PRINT_LEVEL_INFO1,
-			"vdev_id %d bundle_require %d tx_bytes %ld time_in_ms %d high_th %d low_th %d\n",
+			"vdev_id %d bundle_require %d tx_bytes %ld time_in_ms %d high_th %lld low_th %lld\n",
 			vdev->vdev_id, vdev->bundling_reqired, tx_bytes,
 			time_in_ms, high_th, low_th);
 }
